Modify: handle getProfile endpoint in friend, apply creator type change
