+++ routes/api/friend.js
@@ -1,11 +1,10 @@
 const express = require("express");
 const router = express.Router();
 const friendController = require("../controller/friend");
-const indexController = require("../controller/index");
 
 router.get("/", friendController.getFriends);
 
-router.get("/:id", indexController.getProfile);
+router.get("/:id", friendController.getProfile);
 
 router.post("/new", friendController.sendFriendRequest);
