+++ routes/controller/friend.js
@@ -2,6 +2,13 @@ const createError = require("http-errors");
 const mongoose = require("mongoose");
 
 const User = require("../../models/User");
+const Activity = require("../../models/Activity");
+const Sleep = require("../../models/Sleep");
+const Meal = require("../../models/Meal");
+const Album = require("../../models/CustomAlbum");
+const Grid = require("../../models/CustomGrid");
+
+const getPastISOTime = require("../utils/getPastISOTime");
 const ACCESS_LEVELS = require("../../constants/accessLevels");
 const { ERROR } = require("../../constants/messages");
 const {
@@ -10,6 +17,66 @@ const {
 
 async function getFriends(req, res, next) {
   //
+}
+
+async function getProfile(req, res, next) {
+  try {
+    const { creator } = req;
+
+    if (req.accessLevel !== ACCESS_LEVELS.CREATOR) {
+      throw createError(UNAUTHORIZED);
+    }
+
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+      throw createError(BAD_REQUEST, ERROR.INVALID_FRIEND_ID);
+    }
+
+    const friendId = mongoose.Types.ObjectId(req.params.id);
+
+    if (!creator.friends.includes(friendId)) {
+      throw createError(UNAUTHORIZED);
+    }
+
+    const friend = await User.findById(friendId);
+
+    if (!friend) {
+      next(createError(NOT_FOUND));
+    }
+
+    const {
+      stroke, scores, lastAccessDate, profileUrl, name,
+    } = friend;
+    const now = new Date();
+    const { pastAMonthAgo } = getPastISOTime(now);
+
+    const matchOption = { friendId, date: { $gte: pastAMonthAgo, $lte: now } };
+
+    const activity = Activity.aggregate([{ $match: matchOption }, { $addFields: { category: "activity" } }]);
+    const meal = Meal.aggregate([{ $match: matchOption }, { $addFields: { category: "meal" } }]);
+    const sleep = Sleep.aggregate([{ $match: matchOption }, { $addFields: { category: "sleep" } }]);
+    const album = Album.aggregate([{ $match: matchOption }, { $addFields: { type: "album" } }]);
+    const grid = Grid.aggregate([{ $match: matchOption }, { $addFields: { type: "grid" } }]);
+
+    const recentDataPerModel = await Promise.all([
+      activity, meal, sleep, album, grid,
+    ]);
+
+    const data = recentDataPerModel.reduce((data, dataPerModel) => {
+      return data.concat(dataPerModel);
+    }, []).sort((a, b) => b.date - a.date);
+
+    res.status(OK);
+    res.json({
+      stroke,
+      scores,
+      lastAccessDate,
+      profileUrl,
+      name,
+      data,
+    });
+  } catch (err) {
+    next(err);
+  }
 }
 
 async function sendFriendRequest(req, res, next) {
@@ -58,6 +125,7 @@ async function deleteFriendDetail(req, res, next) {
 
 module.exports = {
   getFriends,
+  getProfile,
   sendFriendRequest,
   patchFriendDetail,
   deleteFriendDetail,
